Creation of the BaseRepository implementation Application  of the baserepository implementation for appservice creation of the servicecollection class for repository and appservice creation of prediction domain model

+++ SleekPredictionPunter.DataInfrastructure/PredictionDbContext.cs
@@ -46,7 +46,7 @@ namespace SleekPredictionPunter.DataInfrastructure
 			return base.SaveChanges();
 		}
 
-		public async Task<int> SaveChangesAsync()
+		public async Task<long> SaveChangesAsync()
 		{
 			Audit();
 			return await base.SaveChangesAsync();
